[cr] cleanup geometry branch
